+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [Released]
 
 ## [1.0.27] - 2023-03-19
-- Function to_json was renamed to function to_h because its name is more appropriate.
+- Function to_json was renamed to function as_json because its name is more appropriate. Function to_json was added.
 
 ## [1.0.26] - 2023-03-03
 - Custom result attribute was added to Result model.